# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KostIn API",
        "description": "Backend API documentation for KostIn marketplace kos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API server"
        }
    ],
    "paths": {
...
        "/api/v1/properties/{id}": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Show property",
                "operationId": "8c58403de53ed9fd2db94661ebb0ccdc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Property detail"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Update property",
                "operationId": "0c1bfc464d4281debbee4491d6ebdc52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated property"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Properties"
                ],
                "summary": "Delete property",
                "operationId": "1e3ea6dd2f1d3f4b3dab113ea720fa3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/api/v1/rooms": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "List rooms",
                "operationId": "31a8839ec4012ed828bcc67b19738698",
                "parameters": [
                    {
                        "name": "room_type_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of rooms"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Create room",
                "operationId": "b660449fe16228d1c90a5b92f5716e7e",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/v1/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Show room",
                "operationId": "20e96d2b586955bc2a4d81fdbff8789d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room detail"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Update room",
                "operationId": "31f55fd305dcdddec286f6ab7b6a3ee2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated room"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Delete room",
                "operationId": "7518052060eb6c801b34f9c0de418db6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}